Add steps to setup display and buttons on Pirate Audio HAT
* Make script more robust so it can be executed multiple times without
breaking something.
* Improve README.md

+++ components/audio/PirateAudioHAT/setup_pirateAudioHAT.sh
@@ -19,17 +19,25 @@ sudo systemctl stop phoniebox-gpio-buttons.service
 sudo systemctl disable phoniebox-gpio-buttons.service
 
 printf "Adding settings to /boot/config.txt...\n"
-sudo cp /boot/config.txt /boot/config.txt.bak
+if [[ ! -f /boot/config.txt.bak ]]; then
+    sudo cp /boot/config.txt /boot/config.txt.bak
+fi
 
-echo "gpio=25=op,dh" | sudo tee -a /boot/config.txt > /dev/null
-echo "dtoverlay=hifiberry-dac" | sudo tee -a /boot/config.txt > /dev/null
+# Only add the two lines, if they do not exist already
+if ! sudo grep -qe "gpio=25=op,dh" /boot/config.txt; then
+    echo "gpio=25=op,dh" | sudo tee -a /boot/config.txt > /dev/null
+fi
+if ! sudo grep -qe "dtoverlay=hifiberry-dac" /boot/config.txt; then
+    echo "dtoverlay=hifiberry-dac" | sudo tee -a /boot/config.txt > /dev/null
+fi
 
 printf "Adding settings to /etc/asound.conf...\n"
 # Create backup of /etc/asound.conf if it already exists
-if [[ -f /etc/asound.conf ]]; then
+if [[ -f /etc/asound.conf && ! -f /etc/asound.conf.bak ]]; then
     sudo cp /etc/asound.conf /etc/asound.conf.bak
 fi
 
+# Do not add, but replace content if file already exists
 sudo tee /etc/asound.conf << EOF > /dev/null
 pcm.hifiberry {
     type            softvol
@@ -48,13 +56,13 @@ ctl.!default {
 EOF
 
 # Create backup of /etc/mpd.conf if it already exists
-if [[ -f /etc/mpd.conf ]]; then
+if [[ -f /etc/mpd.conf && ! -f /etc/mpd.conf.bak ]]; then
     sudo cp /etc/mpd.conf /etc/mpd.conf.bak
 fi
 
 printf "Add hifiberry as audio_output in /etc/mpd.conf...\n"
-
-if ! grep -qe "HiFiBerry DAC+ Lite" /etc/mpd.conf; then
+# Only add, if it does not exist already
+if ! sudo grep -qe "HiFiBerry DAC+ Lite" /etc/mpd.conf; then
     sudo sed -i "/# An example of an ALSA output:/ r /dev/stdin" /etc/mpd.conf <<'EOG'
 audio_output {
         enabled         "yes"
@@ -74,6 +82,38 @@ fi
 printf "Set mixer_control name in /etc/mpd.conf...\n"
 mixer_control_name="Master"
 sudo sed -i -E "s/^(\s*mixer_control\s*\")[^\"]*(\"\s*# optional)/\1\\${mixer_control_name}\2/" /etc/mpd.conf
+
+printf "Activating SPI...\n"
+sudo raspi-config nonint do_spi 0
+
+printf "Installing Python dependencies...\n"
+sudo apt-get -y -qq install python3-pil python3-numpy
+
+printf "Installing mopidy plugins...\n"
+sudo pip3 --quiet install Mopidy-PiDi pidi-display-pil pidi-display-st7789 mopidy-raspberry-gpio
+
+# Only add, if it does not exist already
+printf "Editing mopidy configuration...\n"
+if ! sudo grep -qe "raspberry-gpio" /etc/mopidy/mopidy.conf; then
+    sudo tee -a /etc/mopidy/mopidy.conf << EOH > /dev/null
+
+[raspberry-gpio]
+enabled = true
+bcm5 = play_pause,active_low,150
+bcm6 = volume_down,active_low,150
+bcm16 = next,active_low,150
+bcm20 = volume_up,active_low,150
+
+[pidi]
+enabled = true
+display = st7789
+EOH
+else
+    printf "/etc/mopidy/mopidy.conf is already configured. Skipping...\n"
+fi
+
+printf "Enable access for modipy user...\n"
+sudo usermod -a -G spi,i2c,gpio,video mopidy
 
 printf "You need to reboot to apply the settings.\n"
 question "Do you want to reboot now"
